Add experimental unity build option to CMake configuration
