+++ CMakeLists.txt
@@ -331,6 +331,17 @@ else()
   )
 endif()
 
+hpx_option(
+  HPX_WITH_UNITY_BUILD
+  BOOL
+  "Enable unity build for certain build targets (experimental, requires CMake 3.16 or newer) (default OFF)"
+  OFF
+  ADVANCED
+)
+if(HPX_WITH_UNITY_BUILD AND (CMAKE_VERSION VERSION_LESS "3.16"))
+  hpx_error("HPX_WITH_UNITY_BUILD=ON but this requires CMake 3.16 or newer.")
+endif()
+
 # ##############################################################################
 # Some platforms do not support dynamic linking. Enable this to link all
 # libraries statically. This also changes some of the internals of HPX related
